feat: Adding the model and metrics to the project.
The added model is an EfficientNet with frozen weights but unfrozen head weights.
We simply have a linear dense layer in the classification head. Thus, no fine-
tuning is applied. In future version, this task will be considered. The F1-score
with the macro argument was also chosen as validation monitored metric. We come
to select this metric since it is adapted for imbalanced dataset, such as ours
and can easily reflect the difference between the predicted label and the GT.

+++ src/data/dataset.py
@@ -80,7 +80,7 @@ class BoneAgeDataset(Dataset):
                 tensor_image: torch.Tensor = self.final_transform(image=processed_file)['image']
 
         tensor_label: torch.Tensor = torch.tensor(label, dtype=torch.long)
-        tensor_boneage: torch.Tensor = torch.tensor(boneage, dtype=torch.long) 
+        tensor_boneage: torch.Tensor = torch.tensor(boneage, dtype=torch.float32) 
         tensor_male: torch.Tensor = torch.tensor(male, dtype=torch.float32)
         
         return tensor_image, tensor_label, tensor_boneage, tensor_male
